+++ CHANGELOG.md
@@ -12,7 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - `Replicated` is no longer automatically inserted on clients, only `Remote`. `scene::replicate_into` will serialize all entities that have either `Remote` or `Replicated`.
+- `DeferredEntity::flush` now consumes the entity.
 - Rename `scene` module and feature into `world_serialization`.
+- Rename `DeferredChanges` into `EntityScratch`.
 
 ## [0.39.4] - 2026-04-08